Only run cleanup once a minute instead of once per second
This should hopefully still be often enough. In a small cache, the
cleanup run takes 76 millisecond, and if that is run once per second
(as it does in a build with less than 100% cache hit rate), it will
spend 7% of its execution time on sorting the cache index in the
cleanup function.