Fix cleanup when using the filedb storage backend
Previously, this didn't clean anything up at all, since totallen
just ended up as NaN (since row.len was undefined and added to
totallen).

Use the right field, 'lastuse', instead of 'lru'. Make the comparator
function return an integer (if b is larger than, equal to or smaller
than a), returning the items sorted in the right order.

Use smaller than or equal when comparing lastuse, to actually remove
the first item that should be removed, with lastuse equal to the
chosen threshold.